Kims changes in R added manually here
The stupid stupid personal token resetting makes me so mad and it makes me confused and i would have to restart R i think so instead I am copying and pasting this stuff. Main points are the addition of the histogram and the testing_labels changes to Youssef's stuff. Also ran a stupid LSTM model that took a very frustrating long time. urg